[feature] Process only needed recommendations (#11)

+++ src/engines/link_between_items.py
@@ -14,6 +14,8 @@ import numpy as np
 class LinkBetweenItems(Engine):
 
     def train(self):
+        if self.check_if_necessary() is False:
+            return
         with db as session:
             session.execute(
                 text('DELETE FROM "%s" WHERE content_type0 <> content_type1' % (self.__media__[0].tablename_similars)))
@@ -88,3 +90,30 @@ class LinkBetweenItems(Engine):
 
                 self.logger.info("%s + %s similarity reloading performed in %s (%s lines)" %
                                  (m.content_type, m.other_content_cmp[i], datetime.utcnow()-st_time, len(values)))
+            self.store_date(m.content_type)
+
+    def check_if_necessary(self):
+        for media in self.__media__:
+            if media.content_type in [
+                ContentType.APPLICATION,
+                ContentType.BOOK
+            ]:
+                continue
+
+            df = pd.read_sql_query(
+                'SELECT last_launch_date FROM "engine" WHERE engine = \'%s\' AND content_type = \'%s\'' % (self.__class__.__name__, str(media.content_type).upper()), con=db.engine)
+
+            if df.shape[0] == 0:
+                # means that this engine has never been launched.
+                return True
+
+            last_launch_date = df.iloc[0]["last_launch_date"]
+
+            df = pd.read_sql_query(
+                'SELECT COUNT(*) AS c FROM "%s_added_event" WHERE occured_at > \'%s\'' % (media.content_type, last_launch_date), con=db.engine)
+
+            if df.iloc[0]["c"] != 0:
+                # New change occured
+                return True
+
+        return False
